valiadtion erroe in signup page

+++ components/RestaurantSignUp.tsx
@@ -4,6 +4,8 @@ import { useState } from 'react';
 import toast from 'react-hot-toast';
 
 export default function RestaurantSignUp() {
+    const [error, seterror] = useState(false)
+
     const router = useRouter();
     const [formData, setFormData] = useState({
         email: '',
@@ -14,6 +16,7 @@ export default function RestaurantSignUp() {
         address: '',
         contact: ''
     });
+    
 
     // Function to handle form input changes
     const handleChange = (e: any) => {
@@ -24,36 +27,46 @@ export default function RestaurantSignUp() {
         }));
     };
 
+
     // Function to handle form submission
     const handleSubmit = async (e: any) => {
         e.preventDefault();
 
+        const emptyFields = Object.values(formData).some(value => value.trim() === '');
+        if (emptyFields) {
+            toast.error("All fields are required !!");
+            return;
+        }
+    
+        // Check if password and confirmPassword match
+        if (formData.password !== formData.confirmPassword) {
+            toast.error("Passwords do not match");
+            return;
+        }
+
+   
+
+          // Perform sign-up process if all the validation pass
         const result = await ragistre(formData);
-
-
-        console.log(result);
         delete result.password;
         localStorage.setItem("restaurantUser", JSON.stringify(result));
         toast.success('User is Registered')
-
-        try {
-            //  const result = await ragistre(formData)
-            setFormData({
-                email: '',
-                password: '',
-                confirmPassword: '',
-                name: '',
-                city: '',
-                address: '',
-                contact: ''
-            })
-            
-            router.push("/restaurant/dashbord")
-        } catch (error) {
-            console.log(error);
-            console.log("user is not ragister")
-        }
+        router.push("/restaurant/dashbord")
     };
+
+    const  handlereset =()=>{
+        setFormData({
+            name:"",
+            address:"",
+            city:"",
+            password:"",
+            confirmPassword:"",
+            contact:"",
+            email:"",
+        })
+    }
+
+
     return (
         <div className="flex min-h-full flex-1 flex-col justify-center px-6 py-12 lg:px-8">
             <div className="sm:mx-auto sm:w-full sm:max-w-sm">
@@ -75,10 +88,11 @@ export default function RestaurantSignUp() {
                                 name="email"
                                 type="email"
                                 autoComplete="email"
-                                required
+                                
                                 value={formData.email}
                                 onChange={handleChange}
                                 className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                                
                             />
                         </div>
                     </div>
@@ -94,7 +108,7 @@ export default function RestaurantSignUp() {
                                 name="name"
                                 type="text"
                                 autoComplete="name"
-                                required
+                              
                                 value={formData.name}
                                 onChange={handleChange}
                                 className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
@@ -113,7 +127,7 @@ export default function RestaurantSignUp() {
                                 name="password"
                                 type="password"
                                 autoComplete="new-password"
-                                required
+                                
                                 value={formData.password}
                                 onChange={handleChange}
                                 className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
@@ -132,7 +146,7 @@ export default function RestaurantSignUp() {
                                 name="confirmPassword"
                                 type="password"
                                 autoComplete="new-password"
-                                required
+                                
                                 value={formData.confirmPassword}
                                 onChange={handleChange}
                                 className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
@@ -151,7 +165,7 @@ export default function RestaurantSignUp() {
                                 name="city"
                                 type="text"
                                 autoComplete="city"
-                                required
+                                
                                 value={formData.city}
                                 onChange={handleChange}
                                 className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
@@ -170,7 +184,7 @@ export default function RestaurantSignUp() {
                                 name="address"
                                 type="text"
                                 autoComplete="address"
-                                required
+                                
                                 value={formData.address}
                                 onChange={handleChange}
                                 className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
@@ -189,7 +203,7 @@ export default function RestaurantSignUp() {
                                 name="contact"
                                 type="tel"
                                 autoComplete="contact"
-                                required
+                                
                                 value={formData.contact}
                                 onChange={handleChange}
                                 className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
@@ -204,6 +218,14 @@ export default function RestaurantSignUp() {
                             className="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
                         >
                             Sign up
+                        </button>
+
+                        <button
+                        onClick={handlereset}
+                            type='button'
+                            className="flex w-full mt-3 justify-center rounded-md bg-red-500 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm  focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
+                        >
+                            reset
                         </button>
                     </div>
                 </form>
